+++ tests/unit/statsMonitor.ts
@@ -73,7 +73,7 @@ describe('StatsMonitor', () => {
   describe('StatsMonitor.disableWarnings', () => {
     it(`Should NOT raise warnings when thresholds reached and warnings are disabled`, () => {
       const onWarning = sinon.stub();
-      thresholds[STAT_NAME].maxDuration = 2;
+      thresholds[STAT_NAME].max = 10;
       monitor = new StatsMonitor({ getRTCStats, thresholds });
       monitor.enable({});
 
@@ -90,7 +90,7 @@ describe('StatsMonitor', () => {
   describe('StatsMonitor.enableWarnings', () => {
     it(`Should raise warning after re-enabling warnings`, () => {
       const onWarning = sinon.stub();
-      thresholds[STAT_NAME].maxDuration = 2;
+      thresholds[STAT_NAME].max = 10;
       monitor = new StatsMonitor({ getRTCStats, thresholds });
       monitor.enable({});
 
@@ -263,6 +263,71 @@ describe('StatsMonitor', () => {
 
         return wait().then(() => sinon.assert.notCalled(onWarning));
       });
+    });
+
+    context(`'minStandardDeviation' threshold`, () => {
+      it(`Should raise a warning when 'minStandardDeviation' threshold is reached`, () =>
+        new Promise(async resolve => {
+          const statsMonitor = new StatsMonitor({
+            getRTCStats: async () => ({}),
+            thresholds: { audioInputLevel: { maxDuration: 2, sampleCount: 2 } },
+          });
+
+          statsMonitor.enable({});
+
+          statsMonitor.on('warning', warning => {
+            assert.equal(warning.name, 'audioInputLevel');
+            resolve();
+          });
+
+          for (const volume of [10, 10, 10, 10, 10]) {
+            statsMonitor.addVolumes(volume, volume);
+          }
+          await clock.tickAsync(1000);
+
+          for (const volume of [10, 10, 10, 10, 10]) {
+            statsMonitor.addVolumes(volume, volume);
+          }
+          await clock.tickAsync(1000);
+
+          for (const volume of [10, 10, 10, 10, 10]) {
+            statsMonitor.addVolumes(volume, volume);
+          }
+          await clock.tickAsync(1000);
+        }),
+      );
+
+      it(
+        `Should NOT raise warning when 'maxDuration' threshold is reached but with different stat values`,
+        async () => {
+          const statsMonitor = new StatsMonitor({
+            getRTCStats: async () => ({}),
+            thresholds: { audioInputLevel: { maxDuration: 2, sampleCount: 2 } },
+          });
+
+          statsMonitor.enable({});
+
+          const warningHandler: sinon.SinonStub = sinon.stub();
+          statsMonitor.on('warning', warningHandler);
+
+          for (const volume of [10, 150, 300, 400, 500]) {
+            statsMonitor.addVolumes(volume, volume);
+          }
+          await clock.tickAsync(1000);
+
+          for (const volume of [400, 10, 800, 900, 1000]) {
+            statsMonitor.addVolumes(volume, volume);
+          }
+          await clock.tickAsync(1000);
+
+          for (const volume of [50, 100, 1000, 500, 10]) {
+            statsMonitor.addVolumes(volume, volume);
+          }
+          await clock.tickAsync(1000);
+
+          assert(warningHandler.notCalled);
+        },
+      );
     });
 
     context(`'min' and 'max' thresholds`, () => {